requirements.txt
alembic
click
Flask
Flask-Login
Flask-SQLAlchemy
itsdangerous
Jinja2
Mako
MarkupSafe
python-dotenv
sqla-wrapper
SQLAlchemy
Werkzeug
mysql-connector
MySQL
captcha
base64
yagmail